Change RatingScore type from double to float
Updated the `RatingScore` property in the `Movie` class
from `double` to `float` to optimize storage and processing.
This change may impact precision and memory usage.

+++ Cinema.DAL/Entities/Movie.cs
@@ -6,7 +6,7 @@ public class Movie : IEntity
     public int Id { get; set; }
     public string Name { get; set; } = null!;
     public string? Description { get; set; }
-    public double RatingScore { get; set; }
+    public float RatingScore { get; set; }
     public TimeSpan Duration { get; set; }
     public DateTime? ReleaseDate { get; set; }
     public string PosterUrl { get; set; } = null!;
